refactor: SouthPane into functional component (#13676)
